requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
opencv-python